requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.1
python-decouple==3.3
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
